+++ lib/common.js
@@ -22,7 +22,7 @@ module.exports.isValidDomain = function (domain) {
   return '';
 };
 
-module.exports.tplConfigDir = function merge(configDir, defaults) {
+module.exports.tplConfigDir = function (configDir, defaults) {
   var homedir = require('homedir')();
   Object.keys(defaults).forEach(function (key) {
     if ('string' === typeof defaults[key]) {
@@ -32,12 +32,17 @@ module.exports.tplConfigDir = function merge(configDir, defaults) {
   });
 };
 
-module.exports.merge = function merge(defaults, args) {
+module.exports.merge = function (/*defaults, args*/) {
+  var allDefaults = Array.prototype.slice.apply(arguments);
+  var args = args.shift();
   var copy = {};
 
-  Object.keys(defaults).forEach(function (key) {
-    copy[key] = defaults[key];
+  allDefaults.forEach(function (defaults) {
+    Object.keys(defaults).forEach(function (key) {
+      copy[key] = defaults[key];
+    });
   });
+
   Object.keys(args).forEach(function (key) {
     copy[key] = args[key];
   });
@@ -45,7 +50,19 @@ module.exports.merge = function merge(defaults, args) {
   return copy;
 };
 
-module.exports.tplCopy = function merge(copy) {
+module.exports.tplCopy = function (copy) {
+  var url = require('url');
+  var acmeLocation = url.parse(copy.server);
+  var acmeHostpath = path.join(acmeLocation.hostname, acmeLocation.pathname);
+  copy.accountsDir = copy.accountsDir || path.join(copy.configDir, 'accounts', acmeHostpath);
+  // TODO move these defaults elsewhere?
+  //args.renewalDir = args.renewalDir || ':config/renewal/';
+  args.renewalPath = args.renewalPath || ':config/renewal/:hostname.conf';
+  // Note: the /directory is part of the server url and, as such, bleeds into the pathname
+  // So :config/accounts/:server/directory is *incorrect*, but the following *is* correct:
+  args.accountsDir = args.accountsDir || ':config/accounts/:server';
+  hargs.renewalDir = hargs.renewalDir || ':config/renewal/';
+  copy.renewalPath = copy.renewalPath || path.join(copy.configDir, 'renewal', copy.domains[0] + '.conf');
   var homedir = require('homedir')();
   var tpls = {
     hostname: (copy.domains || [])[0]
@@ -70,56 +87,4 @@ module.exports.tplCopy = function merge(copy) {
   });
 
   //return copy;
-};
-
-module.exports.fetchFromDisk = function (args) {
-  // TODO NO HARD-CODED DEFAULTS
-  if (!args.fullchainPath || !args.privkeyPath || !args.certPath || !args.chainPath) {
-    console.warn("missing one or more of args.privkeyPath, args.fullchainPath, args.certPath, args.chainPath");
-    console.warn("hard-coded conventional pathnames were for debugging and are not a stable part of the API");
-  }
-
-  //, fs.readFileAsync(fullchainPath, 'ascii')
-  // note: if this ^^ gets added back in, the arrays below must change
-  return PromiseA.all([
-    fs.readFileAsync(args.privkeyPath, 'ascii')   // 0
-  , fs.readFileAsync(args.certPath, 'ascii')      // 1
-  , fs.readFileAsync(args.chainPath, 'ascii')     // 2
-
-    // stat the file, not the link
-  , fs.statAsync(args.certPath)                   // 3
-  ]).then(function (arr) {
-    var cert = arr[1];
-    var getCertInfo = require('./cert-info').getCertInfo;
-
-    // XXX Note: Parsing the certificate info comes at a great cost (~500kb)
-    var certInfo = getCertInfo(cert);
-
-    return {
-      key: arr[0]                           // privkey.pem
-    , privkey: arr[0]                       // privkey.pem
-
-    , fullchain: arr[1] + '\n' + arr[2]     // fullchain.pem
-    , cert: cert                            // cert.pem
-
-    , chain: arr[2]                         // chain.pem
-    , ca: arr[2]                            // chain.pem
-
-    , privkeyPath: args.privkeyPath
-    , fullchainPath: args.fullchainPath
-    , certPath: args.certPath
-    , chainPath: args.chainPath
-
-    //, issuedAt: arr[3].mtime.valueOf()
-    , issuedAt: Date(certInfo.notBefore.value).valueOf() // Date.now()
-    , expiresAt: Date(certInfo.notAfter.value).valueOf()
-    , lifetime: args.lifetime
-    };
-  }, function (err) {
-    if (args.debug) {
-      console.error("[letsencrypt/lib/common.js] fetchFromDisk");
-      console.error(err.stack);
-    }
-    return null;
-  });
 };
