+++ lib/default-handlers.js
@@ -3,11 +3,11 @@
 var fs = require('fs');
 var path = require('path');
 
-module.exports.agreeToTerms = function (args, agree) {
-  agree(null, args.agreeTos);
+module.exports.agreeToTerms = function (args, agreeCb) {
+  agreeCb(null, args.agreeTos);
 };
 
-module.exports.setChallenge = function (args, challengePath, keyAuthorization, done) {
+module.exports.setChallenge = function (args, domain, challengePath, keyAuthorization, done) {
   //var hostname = args.domains[0];
   var mkdirp = require('mkdirp');
 
@@ -26,14 +26,14 @@ module.exports.setChallenge = function (args, challengePath, keyAuthorization, d
   });
 };
 
-module.exports.getChallenge = function (args, key, done) {
+module.exports.getChallenge = function (args, domain, key, done) {
   //var hostname = args.domains[0];
 
   //console.log("getting the challenge", args, key);
   fs.readFile(path.join(args.webrootPath, key), 'utf8', done);
 };
 
-module.exports.removeChallenge = function (args, key, done) {
+module.exports.removeChallenge = function (args, domain, key, done) {
   //var hostname = args.domains[0];
 
   fs.unlink(path.join(args.webrootPath, key), done);
